Update dependencies
TODO: The new version of hic-straw (now imported as 'hicstraw') has some
nice new features that can replace a good chunk of our code, so it will be
nice to replace that at some point.

The new features look really nice, so I felt bad about my previous
snarking in the comments, and removed it.