Fix incorrect usage of threshold value
The units for the threshold setting when used for reading contacts from
a Hi-C file and when used for creating contacts from a LAMMPS dump are
actually incompatible with eachother. I suppose the fact that they
were the same value in the original notebook was just a coincidence, oops!

This commit seperates the two into different settings, 'count_threshold'
and 'distance_threshold'.